Create a FieldOption for the minimum lifetime of a field.
This is intended for idempotency keys. I made the field generic rather than specific to idempotency keys, but I'm not sure this is right...perhaps it would be easier to migrate from an idempotency key-specific option to a more generic one later than vice versa?

Also, I named the file field_behavior.proto because it is similar in spirit to google/api/field_behavior.proto, but I recognize this may be more confusing than it's worth (especially since I didn't define a FieldBehavior message - I also considered that, but Google's is enum-typed rather than message-typed, so *that* correspondence seemed mroe confusing than useful).

Open to feedback on all of this.