Upgrade buf yaml config and CI (#14)
This PR upgrades the config for buf modules, fixing lint checks, and
updating the CI to push the main module. The config file is upgraded to
v2 (follows this
[guide](https://buf.build/docs/migration-guides/migrate-v2-config-files)).
Ran `buf lint` and corrected the CEL expression on the `Interval` type.
Uses [buf-action](https://github.com/bufbuild/buf-action) config to run
build, checks (lint, format, breaking), and push on each feature branch
and main to the BSR. Please see commit-by-commit for a breakdown of the
steps.

There are a couple of behaviour changes in this PR. This implements the
[default
behaviour](https://github.com/bufbuild/buf-action?tab=readme-ov-file#default-behavior)
as described in the buf-action repo. Every push will now sync with the
BSR. On pull requests checks will be run. On merging to main the deleted
branch will be archived and the latest on main synced.