feat(operations): add the operations component (#18)
This is a fixed up version of #8, introducing
the longrunning.Operation type and service.

The longrunning.Operation has the following changes in contrast to the
Google variant:

1. Using the field name `path` for the resource identifer, rather than
`name`.
2. Using the aep.api.ProblemDetails for the error type, rather than
google.status.Code. This is to ensure that the schema of the JSON LRO
and the Proto LRO are the same, as otherwise proto status codes would be
introduced into JSON APIs.