feat(field): add resource reference syntax (#33)
AEPs do not currently have a first-class annotation for resource
references.

It would be good to have one long-term, as it helps if it is necessary
to extent it beyond existing proto annotations. In addition, there is no
such concept in oas.

At minimum, the new syntax allows for repeated resource reference types,
which will enable patterns where a single field can be populated with
more than one type.